update test structure, initial uuid
